Moved all isolation functions into dedicated isolation_functions sub-package. The isolation_criteria module has been deleted from the repo
