+++ halotools/mock_observables/isolation_functions/__init__.py
@@ -5,4 +5,5 @@ isolation criteria functionality.
 from .spherical_isolation import spherical_isolation
 from .cylindrical_isolation import cylindrical_isolation
 from .conditional_spherical_isolation import conditional_spherical_isolation
+from .conditional_cylindrical_isolation import conditional_cylindrical_isolation
