+++ halotools/mock_observables/isolation_functions/conditional_cylindrical_isolation.py
@@ -8,14 +8,16 @@ import numpy as np
 from functools import partial 
 import multiprocessing 
 
-from .pair_counters.rectangular_mesh import RectangularDoubleMesh
-from .pair_counters.marked_cpairs import (
-    marked_spherical_isolation_engine, marked_cylindrical_isolation_engine)
-from .pair_counters.mesh_helpers import (
+from .cylindrical_isolation import _cylindrical_isolation_process_args
+from .conditional_spherical_isolation import _conditional_isolation_process_weights
+
+from ..pair_counters.rectangular_mesh import RectangularDoubleMesh
+from ..pair_counters.marked_cpairs import marked_cylindrical_isolation_engine
+from ..pair_counters.mesh_helpers import (
     _set_approximate_cell_sizes, _cell1_parallelization_indices, _enclose_in_box)
 
-from ..utils.array_utils import convert_to_ndarray, custom_len
-from ..custom_exceptions import HalotoolsError
+from ...utils.array_utils import convert_to_ndarray, custom_len
+from ...custom_exceptions import HalotoolsError
 
 __all__ = ('conditional_cylindrical_isolation', )
 
@@ -296,292 +298,4 @@ def conditional_cylindrical_isolation(sample1, sample2, rp_max, pi_max,
     is_isolated = np.array(counts, dtype=bool)
     
     return is_isolated
-
-
-def _cylindrical_isolation_process_args(data1, data2, rp_max, pi_max, period, 
-    num_threads, approx_cell1_size, approx_cell2_size):
-    """
-    private function to process the arguents for cylindrical isolation functions
-    """
-    if num_threads is not 1:
-        if num_threads=='max':
-            num_threads = multiprocessing.cpu_count()
-        if not isinstance(num_threads,int):
-            msg = "Input ``num_threads`` argument must be an integer or the string 'max'"
-            raise ValueError(msg)
-    
-    # Passively enforce that we are working with ndarrays
-    x1 = data1[:,0]
-    y1 = data1[:,1]
-    z1 = data1[:,2]
-    x2 = data2[:,0]
-    y2 = data2[:,1]
-    z2 = data2[:,2]
-    
-    N1 = len(x1)
-    
-    rp_max = convert_to_ndarray(rp_max).astype(float)
-    if len(rp_max) == 1:
-        rp_max = np.array([rp_max[0]]*N1)
-    try:
-        assert np.all(rp_max < np.inf)
-        assert np.all(rp_max > 0)
-    except AssertionError:
-        msg = "Input ``rp_max`` must be an array of bounded positive numbers."
-        raise ValueError(msg)
-    
-    max_rp_max = np.amax(rp_max)
-    
-    pi_max = convert_to_ndarray(pi_max).astype(float)
-    if len(pi_max) == 1:
-        pi_max = np.array([pi_max[0]]*N1)
-    try:
-        assert np.all(pi_max < np.inf)
-        assert np.all(pi_max > 0)
-    except AssertionError:
-        msg = "Input ``pi_max`` must be an array of bounded positive numbers."
-        raise ValueError(msg)
-    
-    max_pi_max = np.amax(pi_max)
-    
-    # Set the boolean value for the PBCs variable
-    if period is None:
-        PBCs = False
-        x1, y1, z1, x2, y2, z2, period = (
-            _enclose_in_box(x1, y1, z1, x2, y2, z2, 
-                min_size=[max_rp_max*3.0,max_rp_max*3.0,max_pi_max*3.0]))
-    else:
-        PBCs = True
-        period = convert_to_ndarray(period).astype(float)
-        if len(period) == 1:
-            period = np.array([period[0]]*3)
-        try:
-            assert np.all(period < np.inf)
-            assert np.all(period > 0)
-        except AssertionError:
-            msg = "Input ``period`` must be a bounded positive number in all dimensions"
-            raise ValueError(msg)
-    
-    try:
-        assert max_rp_max < period[0]/3.
-        assert max_rp_max < period[1]/3.
-        assert max_pi_max < period[2]/3.
-    except AssertionError:
-        msg = ("Input ``rp_max`` and ``pi_max`` must both be less than "
-            "input period in the first two and third dimensions respectively.")
-        raise ValueError(msg)
-    
-    if approx_cell1_size is None:
-        approx_cell1_size = [max_rp_max, max_rp_max, max_pi_max]
-    elif custom_len(approx_cell1_size) == 1:
-        approx_cell1_size = [approx_cell1_size, approx_cell1_size, approx_cell1_size]
-    if approx_cell2_size is None:    
-        approx_cell2_size = [max_rp_max, max_rp_max, max_pi_max]
-    elif custom_len(approx_cell2_size) == 1:
-        approx_cell2_size = [approx_cell2_size, approx_cell2_size, approx_cell2_size]
-    
-    return (x1, y1, z1, x2, y2, z2, 
-        rp_max, max_rp_max, pi_max, max_pi_max, period, num_threads, PBCs, 
-        approx_cell1_size, approx_cell2_size)
-
-
-def _conditional_isolation_process_weights(data1, data2, weights1, weights2, cond_func):
-    """
-    private function to process the arguents for conditional isolation functions
-    """
-    
-    correct_num_weights = _func_signature_int_from_cond_func(cond_func)
-    npts_data1 = np.shape(data1)[0]
-    npts_data2 = np.shape(data2)[0]
-    correct_shape1 = (npts_data1, correct_num_weights)
-    correct_shape2 = (npts_data2, correct_num_weights)
-    
-    ### Process the input weights1
-    _converted_to_2d_from_1d = False
-    # First convert weights1 into a 2-d ndarray
-    if weights1 is None:
-        weights1 = np.ones((npts_data1, 1), dtype = np.float64)
-    else:
-        weights1 = convert_to_ndarray(weights1)
-        weights1 = weights1.astype("float64")
-        if weights1.ndim == 1:
-            _converted_to_2d_from_1d = True
-            npts1 = len(weights1)
-            weights1 = weights1.reshape((npts1, 1))
-        elif weights1.ndim == 2:
-            pass
-        else:
-            ndim1 = weights1.ndim
-            msg = ("\n You must either pass in a 1-D or 2-D array \n"
-                   "for the input `weights1`. Instead, an array of \n"
-                   "dimension %i was received.")
-            raise HalotoolsError(msg % ndim1)
-    
-    npts_weights1 = np.shape(weights1)[0]
-    num_weights1 = np.shape(weights1)[1]
-    # At this point, weights1 is guaranteed to be a 2-d ndarray
-    ### now we check its shape
-    
-    if np.shape(weights1) != correct_shape1:
-        if _converted_to_2d_from_1d is True:
-            msg = ("\n You passed in a 1-D array for `weights1` that \n"
-                   "does not have the correct length. The number of \n"
-                   "points in `data1` = %i, while the number of points \n"
-                   "in your input 1-D `weights1` array = %i")
-            raise HalotoolsError(msg % (npts_data1, npts_weights1))
-        else:
-            msg = ("\n You passed in a 2-D array for `weights1` that \n"
-                   "does not have a consistent shape with `data1`. \n"
-                   "`data1` has length %i. The input value of `cond_func` = %i \n"
-                   "For this value of `cond_func`, there should be %i weights \n"
-                   "per point. The shape of your input `weights1` is (%i, %i)\n")
-            raise HalotoolsError(msg % 
-                (npts_data1, cond_func, correct_num_weights, npts_weights1, num_weights1))
-    
-    ### Process the input weights2
-    _converted_to_2d_from_1d = False
-    # Now convert weights2 into a 2-d ndarray
-    if weights2 is None:
-        weights2 = np.ones((npts_data2, 1), dtype = np.float64)
-    else:
-        weights2 = convert_to_ndarray(weights2)
-        weights2 = weights2.astype("float64")
-        if weights2.ndim == 1:
-            _converted_to_2d_from_1d = True
-            npts2 = len(weights2)
-            weights2 = weights2.reshape((npts2, 1))
-        elif weights2.ndim == 2:
-            pass
-        else:
-            ndim2 = weights2.ndim
-            msg = ("\n You must either pass in a 1-D or 2-D array \n"
-                   "for the input `weights2`. Instead, an array of \n"
-                   "dimension %i was received.")
-            raise HalotoolsError(msg % ndim2)
-    
-    npts_weights2 = np.shape(weights2)[0]
-    num_weights2 = np.shape(weights2)[1]
-    # At this point, weights2 is guaranteed to be a 2-d ndarray
-    ### now we check its shape
-    if np.shape(weights2) != correct_shape2:
-        if _converted_to_2d_from_1d is True:
-            msg = ("\n You passed in a 1-D array for `weights2` that \n"
-                   "does not have the correct length. The number of \n"
-                   "points in `data2` = %i, while the number of points \n"
-                   "in your input 1-D `weights2` array = %i")
-            raise HalotoolsError(msg % (npts_data2, npts_weights2))
-        else:
-            msg = ("\n You passed in a 2-D array for `weights2` that \n"
-                   "does not have a consistent shape with `data2`. \n"
-                   "`data2` has length %i. The input value of `cond_func` = %i \n"
-                   "For this value of `cond_func`, there should be %i weights \n"
-                   "per point. The shape of your input `weights2` is (%i, %i)\n")
-            raise HalotoolsError(msg % 
-                (npts_data2, cond_func, correct_num_weights, npts_weights2, num_weights2))
-    
-    return weights1, weights2
-
-
-def _spherical_isolation_process_args(data1, data2, r_max, period, 
-    num_threads, approx_cell1_size, approx_cell2_size):
-    """
-    private function to process the arguents for spherical isolation functions
-    """
-    if num_threads is not 1:
-        if num_threads=='max':
-            num_threads = multiprocessing.cpu_count()
-        if not isinstance(num_threads,int):
-            msg = "Input ``num_threads`` argument must be an integer or the string 'max'"
-            raise ValueError(msg)
-    
-    # Passively enforce that we are working with ndarrays
-    x1 = data1[:,0]
-    y1 = data1[:,1]
-    z1 = data1[:,2]
-    x2 = data2[:,0]
-    y2 = data2[:,1]
-    z2 = data2[:,2]
-    
-    N1 = len(x1)
-    
-    r_max = convert_to_ndarray(r_max).astype(float)
-    if len(r_max) == 1:
-        r_max = np.array([r_max[0]]*N1)
-    try:
-        assert np.all(r_max < np.inf)
-        assert np.all(r_max > 0)
-    except AssertionError:
-        msg = "Input ``r_max`` must be an array of bounded positive numbers."
-        raise ValueError(msg)
-    
-    max_r_max = np.amax(r_max)
-    
-    # Set the boolean value for the PBCs variable
-    if period is None:
-        PBCs = False
-        x1, y1, z1, x2, y2, z2, period = (
-            _enclose_in_box(x1, y1, z1, x2, y2, z2, 
-                min_size=[max_r_max*3.0,max_r_max*3.0,max_r_max*3.0]))
-    else:
-        PBCs = True
-        period = convert_to_ndarray(period).astype(float)
-        if len(period) == 1:
-            period = np.array([period[0]]*3)
-        try:
-            assert np.all(period < np.inf)
-            assert np.all(period > 0)
-        except AssertionError:
-            msg = "Input ``period`` must be a bounded positive number in all dimensions"
-            raise ValueError(msg)
-
-    try:
-        assert max_r_max < period[0]/3.
-        assert max_r_max < period[1]/3.
-        assert max_r_max < period[2]/3.
-    except AssertionError:
-        msg = ("Input ``max(r_max)`` must be less than input period/3 in all dimensions.")
-        raise ValueError(msg)
-
-    if approx_cell1_size is None:
-        approx_cell1_size = [max_r_max, max_r_max, max_r_max]
-    elif custom_len(approx_cell1_size) == 1:
-        approx_cell1_size = [approx_cell1_size, approx_cell1_size, approx_cell1_size]
-    if approx_cell2_size is None:    
-        approx_cell2_size = [max_r_max, max_r_max, max_r_max]
-    elif custom_len(approx_cell2_size) == 1:
-        approx_cell2_size = [approx_cell2_size, approx_cell2_size, approx_cell2_size]
-        
-    return (x1, y1, z1, x2, y2, z2, 
-        r_max, max_r_max, period, num_threads, PBCs, 
-        approx_cell1_size, approx_cell2_size)
-
-
-def _func_signature_int_from_cond_func(cond_func):
-    """
-    Return the function signature available weighting functions. 
-    """
-    
-    if type(cond_func) != int:
-        msg = "\n cond_func parameter must be an integer ID of a weighting function."
-        raise ValueError(msg)
-    
-    if cond_func == 1:
-        return 1
-    elif cond_func == 2:
-        return 1
-    elif cond_func == 3:
-        return 1
-    elif cond_func == 4:
-        return 1
-    elif cond_func == 5:
-        return 2
-    elif cond_func == 6:
-        return 2
-    else:
-        msg = ("The value ``cond_func`` = %i is not recognized")
-        raise HalotoolsError(msg % cond_func)
-
-
-
-
